+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-02-07
+
+### Fixed
+
+- Fixed issue where the user's favorited builds was just returning all builds
+
 ## 2024-02-06
 
 ### Added